nickcolour-horgh: Assign colours to configured nicks only
This updates the nickcolour script to no longer assign random colours to
nicks that are not manually configured to have a colour.

I modernized some of the existing code too, such as how we work with
files.

I removed code no longer needed in this version.

I also improved the command help. And changed color to colour.