sqlquote: Record quote search in two more cases
Recording quote searches is to track quote popularity. Previously we
would not track searches when using 'latest <x>' or 'quote <#>'. Now we
do.

I also move recording search til after we spew the quote to the channel
in the search function. This is so we get the quote out to the channel
ASAP. Recording the search is not as time sensitive.